feat(unit_test): Unit testing for Alarms State.
Mockito is used for mocking dependencies of the providers.

While testing a provider do not forget to override the dependencies of
the provider with the mock, so that you can stub the methods from the
dependencies.

Override the dependencies by passing in a list of the overriden
providers as an argument to `overrides` parameter of
ProviderContainer(). See
[this](https://riverpod.dev/docs/cookbooks/testing#overriding-the-behavior-of-a-provider-during-tests).

Make sure that you do not reuse ProviderContainer between tests. Use the
debugState property of on container.read(<provider>.notifier) to access
the state inside the tests to use in expect.

Remember to use GenerateMocks annotation from mockito or manually create
the correct mocks compliant with null safety. Read more
[here](https://github.com/dart-lang/mockito/blob/master/NULL_SAFETY_README.md).