feat(widget_test): Home View tests. 100% coverage.
When using Future.delayed() inside fake api implementations -> take care
that you call `tester.pump(const Duration(seconds: <delay time>));` before
testing for widgets that are dependent on the fake api implementation.

Riverpod is really nice at letting you override providers during tests.
Reduces a lot of hassle and mocking needs.

I was expecting the screen to load CircularProgressIndicator in the
first frame. But that is not happening. Need to find the reason.